# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "myProperty API",
        "description": "developed by Jonathan",
        "contact": {
            "email": "jonathanx56@hotmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:80",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/v1/real-states": {
            "get": {
                "tags": [
                    "real-states"
                ],
                "summary": "List Real-state",
                "operationId": "list",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "real-states"
                ],
                "summary": "Create Real-state",
                "operationId": "create",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain, application/json"
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "-   Enter the user id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "-   Enter the title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "-   Enter the description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "-   Enter the content",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "-   Enter the price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "-   Enter the slug",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "description": "-   Enter the bedrooms",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "bathrooms",
                        "in": "query",
                        "description": "-   Enter the bathrooms",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "property_area",
                        "in": "query",
                        "description": "-   Enter the property area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "total_property_area",
                        "in": "query",
                        "description": "-   Enter the property total property area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "categories[]",
                        "in": "query",
                        "description": "-   Enter the property categories",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "address_id",
                        "in": "query",
                        "description": "-   Enter the property address",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "images",
                        "in": "formData",
                        "description": "-   Enter the images",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/api/v1/real-states/{id}": {
            "get": {
                "tags": [
                    "real-states"
                ],
                "summary": "Show Real-state",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "real-states"
                ],
                "summary": "Update Real-state",
                "operationId": "update",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "text/plain, application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the real-state id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "-   Enter the user id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "-   Enter the title",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "-   Enter the description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "-   Enter the content",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "-   Enter the price",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "slug",
                        "in": "query",
                        "description": "-   Enter the slug",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "bedrooms",
                        "in": "query",
                        "description": "-   Enter the bedrooms",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "bathrooms",
                        "in": "query",
                        "description": "-   Enter the bathrooms",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "property_area",
                        "in": "query",
                        "description": "-   Enter the property area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "total_property_area",
                        "in": "query",
                        "description": "-   Enter the property total property area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "categories[]",
                        "in": "query",
                        "description": "-   Enter the property categories",
                        "required": true,
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "collectionFormat": "multi"
                    },
                    {
                        "name": "address_id",
                        "in": "query",
                        "description": "-   Enter the property address",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "images",
                        "in": "formData",
                        "description": "-   Enter the images",
                        "required": true,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "real-states"
                ],
                "summary": "Delete Real-state",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "-   Enter the id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}